requirements_computer.txt
opencv-contrib-python>=4.5.3
python-chess>=1.9.4
jinja2>=3.0.1
numpy>=1.19.0
scikit-learn>=0.24.0